Reduce mandatory dependencies by making transport imports optional
This change addresses the issue where importing go-libp2p pulls in a large
set of indirect dependencies even when only a single transport is used.

Changes:
- Moved transport imports from defaults.go to a separate default_transports.go
  file to avoid forcing all transport dependencies when users specify custom
  transports
- Moved QUIC-specific options (QUICReuse) from options.go to options_quic.go
  to avoid importing quicreuse package in the main options file
- Moved QUIC-related code from config/config.go to config/config_quic.go,
  including PrivKeyToStatelessResetKey and PrivKeyToTokenGeneratorKey functions
- Removed direct imports of quic-go and quicreuse from main package files

Benefits:
- Users who only use TCP transport no longer pull in QUIC, WebSocket,
  WebTransport, and WebRTC dependencies
- Eliminates mandatory quic-go dependency (~1.6MB) when QUIC is not used
- Reduces binary size by ~8MB when unused transports are excluded
- Maintains backward compatibility - DefaultTransports still works as before

The transport abstractions no longer leak into the base Options and Config
packages, and users can now use go-libp2p with significantly fewer
dependencies when they only need a subset of transports.

Fixes: Allow using go-libp2p with fewer mandatory dependencies
